Implementacion de la capa de negocio y el metodo GitUserView
