+++ app/src/main/java/com/twelveHours/gitcraft/MainActivity.kt
@@ -7,6 +7,7 @@ import android.widget.Button
 import com.twelveHours.gitcraft.datos.GitHubServiceRequest
 import com.twelveHours.gitcraft.entidad.Repository
 import com.twelveHours.gitcraft.entidad.User
+import com.twelveHours.gitcraft.negocio.GitUserView
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
@@ -28,57 +29,16 @@ class MainActivity : AppCompatActivity() {
         button.setOnClickListener {
             val username = "EduardJiron"
 
-            githubApiService.getUser(username).enqueue(object : Callback<User> {
-                override fun onResponse(call: Call<User>, response: Response<User>) {
-                    if (response.isSuccessful) {
-                        val user = response.body()
-                        val name = user?.User ?: user?.login ?: "Unknown user"
-                        val followers = user?.followers ?: "Unknown followers"
-                        val following = user?.following ?: "Unknown following"
-                        Log.d("GithubApi", "Name: $name")
-                        Log.d("GithubApi", "Followers: $followers")
-                        Log.d("GithubApi", "Following: $following")
-                    } else {
-                        Log.e("GithubApi", "Error: ${response.code()}")
-                    }
-                }
+            val call =GitUserView()
 
-                override fun onFailure(call: Call<User>, t: Throwable) {
-                    Log.e("GithubApi", "Error: ${t.message}")
-                }
-            })
-            githubApiService.getfollo(username).enqueue(object : Callback<List<User>> {
-                override fun onResponse(call: Call<List<User>>, response: Response<List<User>>) {
-                    if (response.isSuccessful) {
-                        val followers = response.body()
-                        followers?.forEach {
-                            Log.d("GithubApi", "Followin: ${it.login}")
-                        }
-                    } else {
-                        Log.e("GithubApi", "Error: ${response.code()}")
-                    }
-                }
 
-                override fun onFailure(call: Call<List<User>>, t: Throwable) {
-                    Log.e("GithubApi", "Error: ${t.message}")
-                }
-            })
-            githubApiService.getStarredRepositories(username).enqueue(object : Callback<List<Repository>> {
-                override fun onResponse(call: Call<List<Repository>>, response: Response<List<Repository>>) {
-                    if (response.isSuccessful) {
-                        val followers = response.body()
-                        followers?.forEach {
-                            Log.d("GithubApi", "Favoritos: ${it.name}")
-                        }
-                    } else {
-                        Log.e("GithubApi", "Error: ${response.code()}")
-                    }
-                }
 
-                override fun onFailure(call: Call<List<Repository>>, t: Throwable) {
-                    Log.e("GithubApi", "Error: ${t.message}")
-                }
-            })
+            button.setOnClickListener(){
+                call.getUser(githubApiService, username)
+            }
+
+
+
         }
 
         }
